+++ backend/src/middleware/tokenReader.ts
@@ -1,3 +1,11 @@
+/**
+ * This middleware prevents anonymous users
+ * from accessing sensitive API endpoints.
+ * The sensitive endpoints must be added to the app
+ * after this middleware.
+ * Endpoints that must be available for anonymous users
+ * must be added before this middleware.
+ */
 import { Request, Response } from 'express';
 import { TokenPayload } from '../types/TokenPayload';
 import { readToken } from './../services/tokenService';
@@ -11,14 +19,6 @@ declare global {
     }
 }
 
-/**
- * This middleware prevents anonymous users
- * from accessing sensitive API endpoints.
- * The sensitive endpoints must be added to the app
- * after this middleware.
- * Endpoints that must be available for anonymous users
- * must be added before this middleware.
- */
 const tokenReader = async (
     req: Request,
     res: Response,
