+++ backend/src/services/validateApiRequest.ts
@@ -1,12 +1,6 @@
-import { ApiRequest } from '../types';
-import { GenerationRequestJsonSchema } from '../types/ApiTypes';
-
-class ValidationError extends SyntaxError {
-    name = 'ValidationError';
-}
-
 /**
- * Backend level validation.
+ * This file handles backend level validation
+ * using the validateApiRequests function.
  * Checks that body is json, and
  * contains the necessary fields for an ApiRequest
  *
@@ -17,6 +11,13 @@ class ValidationError extends SyntaxError {
  * Currently does not check type of fields
  *
  */
+import { ApiRequest } from '../types';
+import { GenerationRequestJsonSchema } from '../types/ApiTypes';
+
+class ValidationError extends SyntaxError {
+    name = 'ValidationError';
+}
+
 // eslint-disable-next-line @typescript-eslint/require-await
 const validateApiRequest = async (body: string): Promise<ApiRequest> => {
     const parsed = GenerationRequestJsonSchema.parse(JSON.parse(body));
@@ -25,12 +26,9 @@ const validateApiRequest = async (body: string): Promise<ApiRequest> => {
         throw new ValidationError(
             `Found ${parsed.contexts.length} contexts instead of 2`
         );
-
+    
     //throw error if output length too high with low input length
-    if (
-        parsed.parameters.outputLength === 5 &&
-        parsed.parameters.inputLength < 4000
-    )
+    if (parsed.parameters.outputLength === 5 && parsed.parameters.inputLength < 4000)
         throw new ValidationError(
             `Invalid output/input length combination (${parsed.parameters.outputLength} and ${parsed.parameters.inputLength})`
         );
