+++ backend/src/tests/api.test.ts
@@ -1,10 +1,22 @@
-import { app } from './../app';
+/**
+ * Tests API validation as well as prompt creation.
+ * Does not test the actual router communication, only errors,
+ * since opening communications during testing is tricky.
+ * However, the dummys own tests should cover for that.
+ */
+
+import { server } from '../index';
 import { createPrompt } from '../services';
 import { TokenPayload } from '../types/TokenPayload';
 import { createToken, createUser } from '../services/userService';
 import supertest from 'supertest';
+import { initializeUsers, getUserToken } from './../services/testService';
 
-const api = supertest(app);
+beforeEach(async () => {
+    await initializeUsers();
+});
+
+const api = supertest(server);
 
 describe('API request validation', () => {
     let token: string;
@@ -79,6 +91,26 @@ describe('API request validation', () => {
             .expect(400);
     });
 });
+
+describe('backend dummy communication, POST /api/textgen', () => {
+    const data = {
+        wrong_data: 'incorrect',
+    };
+
+    test('is not available for anonymous users', async () => {
+        await api.post('/api/textgen').send(data).expect(401);
+    });
+
+    test('router responds with 400 for wrong json request', async () => {
+        const cookie = await getUserToken(api);
+        await api
+            .post('/api/textgen')
+            .set('Cookie', cookie)
+            .send(data)
+            .expect(400);
+    });
+});
+
 
 describe('promptService: Prompt creation', () => {
     test('Returns a prompt with valid parameters, high quality', () => {
