+++ backend/src/tests/error_text.test.ts
@@ -1,3 +1,6 @@
+/**
+ * Tests the error messages related to user log in
+ */
 /* eslint-disable no-console */
 import supertest from 'supertest';
 import { app } from '../app';
@@ -5,9 +8,7 @@ import { initializeUsers } from '../services/testService';
 
 const api = supertest(app);
 
-/**
- * Tests the error messages
- */
+
 
 beforeEach(async () => {
     await initializeUsers();
