+++ backend/src/types/Presets.ts
@@ -1,3 +1,6 @@
+/**
+ * Defines the type for a preset
+ */
 import { Prompt } from './Prompt';
 
 export interface Preset {
